third_party/blink/web_tests/platform/generic/fast/css/resource-timing-for-unloaded-font-expected.txt

Unloaded web fonts should not report Resource Timing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getDocumentFontFaces()[0].status is "unloaded"
PASS entries.length is 0
PASS successfullyParsed is true

TEST COMPLETE
